Moved Outputs object and cleanup.
